Added code to get books of a publisher including authors.
In the PublisherVM View model, added two new view models:
	- BookAuthorVM (Has a bookname and list of book authors)
	- PublisherWithBooksAndAuthorsVM Has a publisher name and a list of BookAuthorVM book authors.

Created new method in PublishersService, the GetPublisherData method.
This method returns the publisher name as well as a list of books with book authors.

In the PublishersController, added a new get endpoint
The GetPublisherData method. This method uses the GetPublisherData method created in the PublishersService