Added two new parameters to the GetAllPublishers method in the publisherService class.
These parameters are the sortBy and the searchString, both of type string.

if sortBy is null, the publishers will be sorted by name in ascending order.
If sortBy is "name_desc", the publishers will be sorted by name in descending order.

The searchString parameter searches for specified characters in the publisher names.

Next will be learning how to do server side paging